% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite_widgets.R
\name{add_counter}
\alias{add_counter}
\title{Add a counter with text}
\usage{
add_counter(my_html, into = "<body>", text, counter_id, value = "")
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{text}{character string; description of the counter.}

\item{counter_id}{character string; id for the span tag which shows the count.}

\item{value}{character string; the counter default value.}
}
\description{
This widget consists of two spans and a slider.
}
