% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_widgets.R
\name{add_dropdown_list}
\alias{add_dropdown_list}
\title{Add drop-down list to body}
\usage{
add_dropdown_list(my_html, into = "<body>", options, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{options}{character vector; the options in the drop-down list. Each item must be unique.}

\item{...}{Extra parameter that goes into a tag, e.g. In <div type = "xxx">, 'type = "xxx"' goes to ... .}
}
\description{
Add drop-down list to body
}
