% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite_widgets.R
\name{add_google_play_pause}
\alias{add_google_play_pause}
\title{Add a google-style play-pause button}
\usage{
add_google_play_pause(my_html, into = "<body>", ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{...}{Other parameters to be passed as tag attributes.}
}
\description{
This widget adds a switch button
}
