% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_header.R
\name{add_google_style}
\alias{add_google_style}
\title{Add google material icon}
\usage{
add_google_style(my_html)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}
}
\description{
Add google material icon
}
