% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite_widgets.R
\name{add_google_style_button}
\alias{add_google_style_button}
\title{Add a google-style single image button}
\usage{
add_google_style_button(my_html, into = "<body>", material_id, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{material_id}{Material id from google material design, e.g. play_circle_outline}

\item{...}{Other parameters to be passed as tag attributes.}
}
\description{
This widget adds a button
}
