% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html.R
\name{add_js_library}
\alias{add_js_library}
\title{Add js libraries src files to html header}
\usage{
add_js_library(my_html, js_libs)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{jslibs}{A vector of strings; JS libraries to use. Currently support
'plotly', 'p5', 'd3', 'vega'.}
}
\description{
Add js libraries src files to html header
}
