% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_body.R
\name{add_script_from_link}
\alias{add_script_from_link}
\title{Add Javascript from link}
\usage{
add_script_from_link(my_html, link)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{link}{weblink to the JS file}
}
\description{
Add Javascript from link
}
