% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html.R
\name{add_slider}
\alias{add_slider}
\title{Add slider to body}
\usage{
add_slider(my_html, type, id, min, max, value, oninput, before, after)
}
\description{
Add slider to body
}
\note{
This function is created using the factory function.
}
