% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite_widgets.R
\name{add_slider_with_text}
\alias{add_slider_with_text}
\title{Add a slider with text description (to body)}
\usage{
add_slider_with_text(my_html, into = "<body>", text, text_id, value,
  oninput = "", input_default = T, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{text}{character string; description of the slider.}

\item{text_id}{character string; id for the span tag which contains the text.}

\item{value}{character string; the slider default value.}

\item{oninput}{character string; command to run when slider updates.}

\item{input_default}{T or F; whether to use the default oninput command (which controls the textbox).}

\item{...}{Extra parameter that goes into a tag, e.g. In <div type = "xxx">, 'type = "xxx"' goes to ... .}
}
\description{
This widget consists of two spans and a slider.
}
