% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_header.R
\name{add_style_from_file}
\alias{add_style_from_file}
\title{Add css style to html header}
\usage{
add_style_from_file(my_html, file)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{file}{character string; path to the css file.}
}
\description{
Add css style to html header
}
