% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vr.R
\name{add_vr_3d_points}
\alias{add_vr_3d_points}
\title{Add VR multiple spheres from dataframe}
\usage{
add_vr_3d_points(my_html, into, df0, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{df0}{A dataframe containing x, y, z coordinates, radius r (optional) and
color (optional). Note that the names of the columns are strict, i.e. must be
'x', 'y', 'z', 'r'(optional), 'color'(optional).}

\item{...}{Extra parameter that goes into a tag, e.g. In <div type = "xxx">, 'type = "xxx"' goes to ... .}
}
\description{
Add VR multiple spheres from dataframe
}
