% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vr.R
\name{add_vr_asset_item}
\alias{add_vr_asset_item}
\title{Add VR asset item}
\usage{
add_vr_asset_item(my_html, into, id, src, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{id}{character string; asset id for future referral.}

\item{src}{character string; path to asset.}

\item{...}{Extra parameter that goes into a tag, e.g. In <div type = "xxx">, 'type = "xxx"' goes to ... .}
}
\description{
Add VR asset item
}
