% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{developing}
\alias{developing}
\title{Helper function to aid developing and testing codes}
\usage{
developing(file, code, time = 1, catch_error = T)
}
\arguments{
\item{file}{character string; the filepath.}

\item{code}{character string; the code to be run when a change is made to file.}

\item{time}{numeric; monitoring change of file every "time" seconds.}

\item{catch_error}{boolean; whether to print out error.}
}
\description{
Helper function to aid developing and testing codes
}
