% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{developing_html}
\alias{developing_html}
\title{Helper function to aid developing and testing html file}
\usage{
developing_html(file, time = 1)
}
\arguments{
\item{file}{character string; the filepath.}

\item{time}{numeric; monitoring change of file every "time" seconds.}
}
\description{
Helper function to aid developing and testing html file
}
