% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{developing_r}
\alias{developing_r}
\title{Helper function to aid developing and testing R codes}
\usage{
developing_r(file, time = 1, catch_error = T)
}
\arguments{
\item{file}{character string; the filepath.}

\item{time}{numeric; monitoring change of file every "time" seconds.}

\item{catch_error}{boolean; whether to print out error.}
}
\description{
Helper function to aid developing and testing R codes
}
