% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html_util.R
\name{get_tag_index}
\alias{get_tag_index}
\title{Get the first index of a tag in a vector of a html string}
\usage{
get_tag_index(tag, my_html)
}
\description{
Get the first index of a tag in a vector of a html string
}
\keyword{internal}
