% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_html5_elements.R
\name{html5_elements}
\alias{html5_elements}
\title{Create builders for html5 elements}
\usage{
html5_elements(my_html, into, insert = T, tag, text, close_tag = T, ...)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}

\item{insert}{T or F, if FALSE, return the element instead of inserting into html.}

\item{tag}{html tag, e.g. div, span, h3.}

\item{text}{character string; the display text on the button}

\item{close_tag}{T or F. Should closing tag be included?}

\item{...}{Other parameter passing to tag attributes.}
}
\description{
A generic function to produce html 5 elements
}
