% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html_util.R
\name{insert_after}
\alias{insert_after}
\title{Insert a script after the tag of a vector of a html string}
\usage{
insert_after(script, tag, my_html)
}
\description{
Insert a script after the tag of a vector of a html string
}
\keyword{internal}
