% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html_util.R
\name{insert_before}
\alias{insert_before}
\title{Insert a script before the tag of a vector of html string}
\usage{
insert_before(script, tag, my_html)
}
\description{
Insert a script before the tag of a vector of html string
}
\keyword{internal}
