% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_html_util.R
\name{insert_into}
\alias{insert_into}
\title{Insert a script into a tag of a html string}
\usage{
insert_into(my_html, script, into)
}
\description{
Insert a script into a tag of a html string
}
\keyword{internal}
