% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{insert_websockets}
\alias{insert_websockets}
\title{Insert a websocket connection and convert html to string.}
\usage{
insert_websockets(filepath, wsUrl)
}
\description{
Insert a websocket connection and convert html to string.
}
\keyword{internal}
