% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_app}
\alias{start_app}
\title{Run an interactive app}
\usage{
start_app(app, host = "localhost", port = 9454, browser = "viewer")
}
\arguments{
\item{app}{An app; output from the 'create_app' function}

\item{host}{character string; Address to host the app.}

\item{port}{integer; Port to host the app.}

\item{browser}{"browser" (web) or "viewer" (R).}
}
\description{
Run an interactive app
}
