% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{tweenr_example}
\alias{tweenr_example}
\title{Function to create example data shown in the tweenr package.}
\usage{
tweenr_example()
}
\description{
Function to create example data shown in the tweenr package.
}
