% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_body.R
\name{add_container}
\alias{add_container}
\title{Add a container (div element) to body}
\usage{
add_container(my_html, id, align, into = "<body>", insert = T)
}
\description{
Add a container (div element) to body
}
