% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_header.R
\name{add_header_title}
\alias{add_header_title}
\title{Add title to html header}
\usage{
add_header_title(my_html, script)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{script}{character string; title.}
}
\description{
Add title to html header
}
