% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_header.R
\name{add_style}
\alias{add_style}
\title{Add css style to html header}
\usage{
add_style(my_html, script)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{script}{character string; css style.}
}
\description{
Add css style to html header
}
