% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_header.R
\name{add_tag_into}
\alias{add_tag_into}
\title{Generic function to add a pair of enclosing "tags" to "script", then insert into "into" of "my_html"}
\usage{
add_tag_into(my_html, script, tag, into)
}
\description{
Generic function to add a pair of enclosing "tags" to "script", then insert into "into" of "my_html"
}
