% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_body.R
\name{add_text}
\alias{add_text}
\title{Add text}
\usage{
add_text(my_html, text, into = "<body>")
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{text}{character string; text to be added.}

\item{into}{character string; unique identifier of a line in the html. Element id is highly recommended.}
}
\description{
Add text
}
