% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_util.R
\name{get_tag_index}
\alias{get_tag_index}
\title{Get the first index of a tag in a vector of a html string}
\usage{
get_tag_index(my_html, tag, after = 0)
}
\description{
Get the first index of a tag in a vector of a html string
}
\keyword{internal}
