% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_util.R
\name{insert_into}
\alias{insert_into}
\title{Generic function to insert a script into a tag of a html string}
\usage{
insert_into(my_html, script, into)
}
\description{
Generic function to insert a script into a tag of a html string
}
\keyword{internal}
