% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{write_html_to_file}
\alias{write_html_to_file}
\title{Convert html to string}
\usage{
write_html_to_file(my_html, file)
}
\arguments{
\item{my_html}{html in a vector of strings; output from 'create_html'.}

\item{file}{Character string; the filepath.}
}
\description{
Convert html to string
}
